pyproject.toml
[tool.poetry]
name = "mentat"
version = "0.1.0"
description = "Mentat is an open-source, fully extensible interface for configuration and deployment of agentic workflows."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
litellm = "^1.55.9"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
